dataview utils
