+++ lib/ui/Dataview.mjs
@@ -57,7 +57,7 @@ A default dataview update method will be assigned if not provided with the datav
 
 The show() and hide() methods will be assigned if not explicit in the dataview configuration. A dataview will have show and hide methods if declared as a [tabview]{@link module:/ui/Tabview~Tabview}.
 
-The dataviewToolbar() method will be called prior to creating the dataview.
+The mapp.ui.utils.dataview.Toolbar() method will be called prior to creating the dataview.
 
 The dataviewMapChange() method will be called after the dataview has been created.
 
@@ -128,7 +128,7 @@ export default async function Dataview(_this) {
   _this.hide ??= hide;
 
   // Create dataview toolbar
-  dataviewToolbar(_this);
+  mapp.ui.utils.dataview.Toolbar(_this);
 
   dataviewMapChange(_this);
 
@@ -304,52 +304,6 @@ async function updateDataview() {
 
   // The setData method is assigned in the create() method.
   typeof this.setData === 'function' && this.setData(response);
-}
-
-/**
-@function dataviewToolbar
-
-@description
-The dataviewToolbar method will create a HTML Element container with target elements for toolbar elements and the dataview itself.
-
-@param {dataview} _this Dataview object.
-@property {Object} _this.toolbar Configuration object for a toolbar supporting the dataview.
-@property {string} _this.dataview The type of dataview
-@property {HTMLElement} _this.target Target element in which the dataview is rendered.
-@property {HTMLElement} [_this.panel] A panel element to be displayed in a tab[view].
-*/
-function dataviewToolbar(_this) {
-  // A toolbar config object is required.
-  if (!_this.toolbar) return;
-
-  // A custome toolbar method has been assigned to the dataview.
-  if (typeof _this.toolbar === 'function') {
-    _this.toolbar();
-  }
-
-  // Target for the dataview.
-  const target = mapp.utils.html.node`<div class="dataview-target">`;
-
-  // Find and execute matching toolbar methods.
-  const toolbarElements = Object.keys(_this.toolbar)
-    .map((key) => mapp.ui.utils[_this.dataview]?.toolbar[key]?.(_this))
-    .filter((item) => !!item);
-
-  // Create an element for the toolbar buttons
-  _this.btnRow = mapp.utils.html
-    .node`<div class="btn-row">${toolbarElements}</div>`;
-
-  // By default btnRow is hidden
-  _this.btnRow.style.setProperty('display', 'none');
-
-  // The panel will be assigned in a tabview.
-  _this.panel = _this.target.appendChild(mapp.utils.html.node`
-    <div class="flex-col">
-      ${_this.btnRow}
-      ${target}`);
-
-  // Assign dataview target as target.
-  _this.target = target;
 }
 
 /**
