+++ lib/ui/utils/_utils.mjs
@@ -56,10 +56,13 @@ import jsonDataview from './jsonDataview.mjs';
 
 import locationCount from './locationCount.mjs';
 
+import dataview from './dataview.mjs';
+
 export default {
   Chart,
   Tabulator,
   cssColourTheme,
+  dataview,
   Json: jsonDataview,
   idleLogout,
   imagePreview,
