+++ lib/ui/utils/jsonDataview.mjs
@@ -13,7 +13,6 @@ export default {
   toolbar: {
     jsonfile,
     csvupload,
-    update,
   },
 };
 
@@ -113,105 +112,4 @@ function csvupload(dataview) {
       }}>`;
 
   return dataview.toolbar.csvupload.input;
-}
-
-function update(dataview) {
-  if (dataview.toolbar.update === true) {
-    dataview.toolbar.update = {};
-  }
-
-  dataview.toolbar.update.button = mapp.utils.html.node`<button 
-    onclick=${() => {
-      const toogle = dataview.toolbar.update.button.classList.toggle('active');
-      if (toogle) {
-        updateDialog(dataview);
-      } else {
-        dataview.toolbar.update.dialog.close();
-      }
-    }}>
-    Update`;
-
-  return dataview.toolbar.update.button;
-}
-
-async function updateDialog(dataview) {
-  const content = mapp.utils.html.node`<div>`;
-
-  const jsonContent = {
-    query: dataview.query,
-    queryparams: dataview.queryparams,
-    data: dataview.data,
-  };
-
-  const jsoneditor = await mapp.ui.elements.jsoneditor({
-    target: content,
-    props: {
-      content: {
-        text: JSON.stringify(jsonContent),
-      },
-      mode: 'text',
-      onRenderMenu: renderUpdateMenu,
-    },
-  });
-
-  dataview.toolbar.update.dialog = {
-    header: 'Dataview Update',
-    content,
-    target: document.getElementById('Map'),
-    closeBtn: true,
-    onClose: (e) => {
-      dataview.toolbar.update.button.classList.remove('active');
-    },
-  };
-
-  mapp.ui.elements.dialog(dataview.toolbar.update.dialog);
-
-  // Create a custom menu for the userLayer jsoneditor control.
-  function renderUpdateMenu(items) {
-    // Push button to add layer to mapview layers.
-    items.push({
-      type: 'button',
-      className: 'material-symbols-outlined-important',
-      text: 'sync',
-      title: 'Update Dataview Query',
-      onClick: updateDataview,
-    });
-
-    items.push({
-      type: 'button',
-      className: 'material-symbols-outlined-important',
-      text: 'data_object',
-      title: 'Set Dataview Data',
-      onClick: setDataview,
-    });
-
-    return items
-      .filter((item) => item.text !== 'table')
-      .filter((item) => item.text !== 'tree')
-      .filter((item) => item.type !== 'separator')
-      .filter((item) => item.className !== 'jse-undo')
-      .filter((item) => item.className !== 'jse-redo')
-      .filter((item) => item.className !== 'jse-search')
-      .filter((item) => item.className !== 'jse-contextmenu')
-      .filter((item) => item.className !== 'jse-sort')
-      .filter((item) => item.className !== 'jse-transform');
-  }
-
-  function updateDataview() {
-    const content = jsoneditor.get();
-
-    const json = JSON.parse(content.text);
-
-    Object.assign(dataview, json);
-
-    dataview.update();
-  }
-
-  function setDataview() {
-    const content = jsoneditor.get();
-
-    const json = JSON.parse(content.text);
-
-    dataview.setData(json.data);
-  }
 }
